+++ R/yyy-deprecate.R
@@ -1,19 +1,19 @@
-vapour_read_geometry_what <- function(dsource, layer = 0L, sql = "", what = "geometry", textformat = "") {
+vapour_read_geometry_what <- function(...) {
   .Defunct("vapour_read_geometry")
 }
-vapour_read_feature_what <- function(dsource, layer = 0L, sql = "", what = "geometry", textformat = "json") {
+vapour_read_feature_what <- function(...) {
   .Defunct("vapour_read_geometry")
 }
 
 
 
-sds_info <- function(x) {
+sds_info <- function(...) {
   .Defunct("vapour_sds_names")
 }
 
-raster_sds_info <- function(x) {
+raster_sds_info <- function(...) {
   .Defunct("vapour_sds_names")
 }
-raster_io <- function(x, band = 1, window, resample = "nearestneighbour", ..., sds = NULL) {
+raster_io <- function(...) {
  .Defunct("vapour_read_raster")
 }